+++ src/App.tsx
@@ -1,7 +1,14 @@
 import { RouterProvider } from 'react-router-dom'
 
+import { ApplicationErrorBoundary } from './components/errors/ApplicationErrorBoundary'
+import { DevelopmentErrorTrigger } from './components/errors/DevelopmentErrorTrigger'
 import { router } from './routes/router'
 
 export function App() {
-  return <RouterProvider router={router} />
+  return (
+    <ApplicationErrorBoundary>
+      <DevelopmentErrorTrigger />
+      <RouterProvider router={router} />
+    </ApplicationErrorBoundary>
+  )
 }